Update CameraPerception.py, final_test.py, and stacking_info.json

+++ final_test.py
@@ -30,26 +30,27 @@ class LetterStacker:
 
     def main(self, letter="H"):
         letter_order = self.letter_list[letter]
-        for row in
-        pass
-
+        for r in range(3):
+            for c in range(3):
+                self.place_block(letter_order[r][c], self.goal_loc[r][c])
 
     def place_block(self, block_color, goal_loc):
         self.img = self.my_camera.frame
+        if block_color == "skip":
+            return False
         while True:
             self.img = self.my_camera.frame
             if self.img is not None:
                 frame = self.img.copy()
-                Frame, coordinates, rect = self.camera_perception.main(frame)
+                Frame, coordinates, rect = self.camera_perception.main(frame, (block_color))
                 cv2.imshow('Frame', Frame)
                 print(f'\n\n MADE IT    Coordinates:  {coordinates} \n\n')
                 print(f'\n\n target loc: {coordinates[0]}, {coordinates[1]}, 1.5 \n\n')
                 self.camera_motion.main(target_loc=(coordinates[0], coordinates[1], 1.5),
-                                   goal_loc=goal_coordinates[target_color], rect=rect)
+                                   goal_loc=goal_loc, rect=rect)
                 break
         self.my_camera.camera_close()
         cv2.destroyAllWindows()
-        pass
 
     def read_json(self):
 
@@ -93,6 +94,7 @@ def write_json(updated_letter_list):
 
 if __name__ == '__main__':
 
-    main()
+    func = LetterStacker()
+    func.main("single_stack")
 
 # print(stacking_info)
